Switching to the new frontend

+++ maps4fsapi/components/task.py
@@ -2,23 +2,28 @@
 
 import os
 
-from fastapi import APIRouter, BackgroundTasks, HTTPException
+from fastapi import APIRouter, BackgroundTasks, HTTPException, Request
 from fastapi.responses import FileResponse
 
 from maps4fsapi.components.models import TaskIdPayload
 from maps4fsapi.config import logger
 from maps4fsapi.limits import dependencies
 from maps4fsapi.storage import Storage
+from maps4fsapi.tasks import TasksQueue
 
 task_router = APIRouter(dependencies=dependencies)
 
 
+@task_router.post("/status")
 @task_router.post("/get")
-def get_task(payload: TaskIdPayload, background_tasks: BackgroundTasks):
+def get_task(payload: TaskIdPayload, request: Request, background_tasks: BackgroundTasks):
     """Retrieve a task result based on the provided task ID.
+    If used the 'status' endpoint, does not return actual data.
+    If used the 'get' endpoint, returns the actual data and removes the entry from storage.
 
     Arguments:
         payload (TaskIdPayload): The payload containing the task ID.
+        request (Request): The request object.
         background_tasks (BackgroundTasks): Background tasks to handle cleanup after response.
     Raises:
         HTTPException: If the task ID is not found, the task failed, or the file is not available.
@@ -28,10 +33,24 @@ def get_task(payload: TaskIdPayload, background_tasks: BackgroundTasks):
     logger.info("Received request to get task with ID: %s", payload.task_id)
     entry = Storage().get_entry(payload.task_id)
     if not entry:
+        # * Order matters! Currently processing task is also in the queue.
+        if TasksQueue().is_processing(payload.task_id):
+            logger.warning("Task ID %s is currently being processed.", payload.task_id)
+            raise HTTPException(
+                status_code=202,
+                detail=f"Task ID {payload.task_id} is currently being processed.",
+            )
+        if TasksQueue().is_in_queue(payload.task_id):
+            logger.warning("Task ID %s is still in the queue.", payload.task_id)
+            raise HTTPException(
+                status_code=204,
+                detail=f"Task ID {payload.task_id} is still in the queue.",
+            )
+
         logger.warning("Task ID %s not found.", payload.task_id)
         raise HTTPException(
             status_code=404,
-            detail="Task ID not found. It's expired or not finished yet.",
+            detail=f"Task ID {payload.task_id} not found. It's expired or not finished yet.",
         )
 
     if not entry.success:
@@ -45,7 +64,7 @@ def get_task(payload: TaskIdPayload, background_tasks: BackgroundTasks):
         logger.warning("No file path found for task ID %s.", payload.task_id)
         raise HTTPException(
             status_code=404,
-            detail="No file path found for the task.",
+            detail=f"No file path found for task ID {payload.task_id}.",
         )
 
     if not os.path.isfile(entry.file_path):
@@ -54,8 +73,16 @@ def get_task(payload: TaskIdPayload, background_tasks: BackgroundTasks):
         )
         raise HTTPException(
             status_code=404,
-            detail="File not found.",
+            detail=f"File not found for task ID {payload.task_id}.",
         )
+
+    endpoint = request.url.path
+    if endpoint.endswith("/status"):
+        return {
+            "success": True,
+            "description": "Task completed successfully. Use the 'get' endpoint to retrieve the file.",
+            "task_id": payload.task_id,
+        }
 
     background_tasks.add_task(Storage().remove_entry, payload.task_id)
     logger.info("Returning file for task ID %s: %s", payload.task_id, entry.file_path)
